requirements-windows.txt
asgiref==3.10.0
attrs==25.4.0
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.11
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==12.0.0
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
